+++ crates/core-new/input_processor.json
@@ -3,8 +3,8 @@
   "proofs": {
     "mmr_meta": {
       "id": 26,
-      "root": "0x758b092f8dc13d2398b3cf7ee5dd72139f888d343ccb6a3ad2ad272feeeb514",
-      "size": 12273695,
+      "root": "0x1fb9e130f89fdca38e93e56d4d60b3cbaf32031f6647170fe21275ac08f09f2",
+      "size": 12276615,
       "peaks": [
         "0x262c4c9b1cb2a036924aecf563dc9952e5f8b41004310adde86f22abb793eb1",
         "0x72f553aac8690d09c95fe6220fdd5a073440631e4ca0a161a92b655d2ac9478",
@@ -13,10 +13,12 @@
         "0x5c4814bbd601bffb5ac9980977a79bf100d4c1ad4f2caa410f7a7c4249a2fd4",
         "0x668035a3620690024dac08a8db46e3316619e4c2a634daaa3175ab16af72deb",
         "0x67cff2a39ca6fb235decefaf5bb63f54c550b97b57e9873751eb9dae35cfcd4",
-        "0x636c962fdb042846fa643a47002fefa1b0279915b203b26f0c01cc8a44c8e12",
-        "0x4ae74f2fc86ab63862fa4a6489ba7878a94559545bfc4c4f9229fc5debe3162",
-        "0x2179946fc2c1a9d150cdffe013b1e2e860438bb069ddce9d4c7d52a8f33df3f",
-        "0x0b3da529a068c564b58520b087782641df0fbfd2b86740ef519423036ae24fd"
+        "0x7c8704b8f5a93be376ac3e3fea4b22fe246e36827f749ae3a1b5b4f4e18348f",
+        "0x14aa5347b888a45c42cc4fe908a69743010594bedaf2e9d1b418979195ba882",
+        "0x47b9f7e6c6f9334536487410025fd95adcf76749746b05a0a6dc6cb8c4ba48f",
+        "0x0283d927018df7111d5d40f91e1660d5575ffe33b64fea839159c54b801d1dc",
+        "0x372ae94c190f23739edfe44a20d1941b6e0910c8785846b1b565d3d766b8c2d",
+        "0x1fdd00942364d582ad4d37c544aad46e684bc0eca2806e4a0d25e029441eb15"
       ]
     },
     "headers": [
