+++ crates/core/src/evaluator/result.rs
@@ -1,14 +1,14 @@
 use anyhow::Result;
 
 use hdp_primitives::processed_types::{
-    account::{ProcessedAccount, ProcessedAccountInFelts},
-    datalake_compute::{ProcessedDatalakeCompute, ProcessedDatalakeComputeInFelts},
-    header::{ProcessedHeader, ProcessedHeaderInFelts},
+    account::ProcessedAccount,
+    cairo_format::{self, IntoFelts},
+    datalake_compute::ProcessedDatalakeCompute,
+    header::ProcessedHeader,
     mmr::MMRMeta,
-    receipt::{ProcessedReceipt, ProcessedReceiptInFelts},
-    storage::{ProcessedStorage, ProcessedStorageInFelts},
-    traits::IntoFelts,
-    transaction::{ProcessedTransaction, ProcessedTransactionInFelts},
+    receipt::ProcessedReceipt,
+    storage::ProcessedStorage,
+    transaction::ProcessedTransaction,
     uint256::Uint256,
 };
 use serde::{Deserialize, Serialize};
@@ -99,11 +99,11 @@ pub struct ProcessedResultFormatted {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub results_root: Option<Uint256>,
     pub tasks_root: Uint256,
-    pub headers: Vec<ProcessedHeaderInFelts>,
+    pub headers: Vec<cairo_format::ProcessedHeader>,
     pub mmr: MMRMeta,
-    accounts: Vec<ProcessedAccountInFelts>,
-    storages: Vec<ProcessedStorageInFelts>,
-    transactions: Vec<ProcessedTransactionInFelts>,
-    transaction_receipts: Vec<ProcessedReceiptInFelts>,
-    pub tasks: Vec<ProcessedDatalakeComputeInFelts>,
+    accounts: Vec<cairo_format::ProcessedAccount>,
+    storages: Vec<cairo_format::ProcessedStorage>,
+    transactions: Vec<cairo_format::ProcessedTransaction>,
+    transaction_receipts: Vec<cairo_format::ProcessedReceipt>,
+    pub tasks: Vec<cairo_format::ProcessedDatalakeCompute>,
 }
