+++ crates/primitives/src/processed_types/cairo_format/mpt.rs
@@ -1,0 +1,38 @@
+use crate::processed_types::mpt::ProcessedMPTProof as BaseProcessedMPTProof;
+
+use super::{felt_vec_unit::FieldElementVectorUnit, traits::IntoFelts};
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+use starknet::core::serde::unsigned_field_element::UfeHex;
+use starknet_crypto::FieldElement;
+
+impl IntoFelts for BaseProcessedMPTProof {
+    type Output = ProcessedMPTProof;
+
+    fn to_felts(&self) -> ProcessedMPTProof {
+        let proof_felts: Vec<FieldElementVectorUnit> = self
+            .proof
+            .iter()
+            .map(|proof| FieldElementVectorUnit::from_hex_str(proof).unwrap())
+            .collect();
+
+        let proof_bytes_len = proof_felts.iter().map(|f| f.bytes_len).collect();
+        let proof_result: Vec<Vec<FieldElement>> =
+            proof_felts.iter().map(|f| f.felts.clone()).collect();
+        ProcessedMPTProof {
+            block_number: self.block_number,
+            proof_bytes_len,
+            proof: proof_result,
+        }
+    }
+}
+
+#[serde_as]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Eq, Hash)]
+pub struct ProcessedMPTProof {
+    pub block_number: u64,
+    /// proof_bytes_len is the byte( 8 bit ) length from each proof string
+    pub proof_bytes_len: Vec<u64>,
+    #[serde_as(as = "Vec<Vec<UfeHex>>")]
+    pub proof: Vec<Vec<FieldElement>>,
+}
