+++ crates/primitives/src/processed_types/mod.rs
@@ -1,12 +1,12 @@
+pub mod cairo_format;
+
 pub mod account;
 pub mod datalake_compute;
-pub mod felt_vec_unit;
 pub mod header;
 pub mod mmr;
 pub mod module;
 pub mod mpt;
 pub mod receipt;
 pub mod storage;
-pub mod traits;
 pub mod transaction;
 pub mod uint256;
